rule34.py: Fix broken get_pool method
The get_pool() method is broken in two ways. First, it is using the API
endpoint URL, when it should be using the interactive website. As a
result, it returns no values. Switch over to the correct URL.

Second, the fast=True parameter causes the method to return a list of
strings, instead of the advertised list of integers. Cast the return
values to ints to fix this.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>