Do not test or advertise the Stats feature
As far as I can tell, the rule34 stats page no longer exists and the
rule34Py.stats feature is totally broken. There doesn't seem to be an
equivalent. So this feature should probably just be deprecated entirely.

Remove this feature from testing, and remove it from the README so that
new users don't think that it works.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>