rule34Py: remove the legacy rule34 client
For some reason the v1.4.0 release included a fork of the client class
module as rule34Py_old.py. It seems to have never been updated and is
not referenced anywhere.

Remove it for code clarity. If devs want to reference the old file, they
can restore it from source-control.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>