rule34Py: do not inherit Exception
It isn't clear to me why, but the rule34Py class inherits from
Exception. I cannot think of a reason that this would be desirable.

Change the rule34Py class inherit from 'Exception' to the implicit
'object'.

Likewise, the client __init__ method sets __isInit__ to False, which is
not used anywhere. Remove that declaration.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>