rule34Py: remove search() deleted param
The search() method is supposed to take a 'deleted' parameter, and
optionally show deleted posts. But that functionality doesn't seem to be
implemented in the upstream API, and the method now just throws a
generic Exception. That's not very helpful.

Instead of always throwing errors, just deprecate that parameter. If it
gets implemented upstream, we can add the functionality back in.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>