rule34Py: deprecate search() ignore_max_limit param
The ignore_max_limit parameter is supposed to allow a user to bypass
checks against a `limit` param that is too large, and pass it to the
upstream API anyway. But there is no point in doing that since, the API
has a hard limit of 1000 and returns that number of results, even if you
request more.

So remove the `ignore_max_limit` parameter.

Also throw a more useful ValueError when limit has an insane value.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>