rule34Py: use a User-Agent from the client
Requests traffic from the Rule34 client uses a hidden module variable
that stores the User-Agent and other headers. Users would prefer to be
able to modify the User-Agent from the client class, and to make generic
HTTP GET requests from the client.

Add a _get() method that wraps the requests.get() method and adds the
client User-Agent as a header. Store the User-Agent in the class.

This commit intentionally doesn't change any client methods to use the
new method.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>