rule34Py: parse the real tagmap
The rule34Py.tagmap() method doens't actually return any tagmap data.
Instead, it retuns the top 100 global tags from the 'toptags' page,
which incidentally does have a tagmap on it which is unparsed.

Instead, implement a class to parse the static /tagmap.html file
javascript to return the actual tagmap data, and return that from the
tagmap() method.

Move the global 100 tags logic to a top_tags() method.

Also regenerate the responses registry.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>