rule34Py: deprecate the version() property
The rule34Py.version() property returns the module version triplet.
Presumably, if a user wanted to learn the module triplet, they could go
to the module-scope variable: `version`, which already returns the same.

Announce that this property will be deprecated in a future release.

Add tests to check that the client gives the deprecation warning, and
that the module-scope variable continues to work as expected.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>