rule34Py: remove the icame() `limit` param
The icame() method pretends to accept a `limit` parameter, which reduces
the number of results in the icame list. But it doesn't actually use this
parameter and always returns 100 values.

Remove this unused parameter. If a user wants to limit their list
results, they can better slice it using native python list operations.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>