rule34Py: make random_post function as website
The rule34Py.random_post() optionally accepts a list of tags. If tags
are provided, the method returns a random post from the first page of
tag search results. Otherwise, it functions like the website Posts >
Random function and returns a random post from the total population.

This alternating behavior is a little strange. The tagged algorithm
isn't really 'random', in the sense that it is limited to the latest
page of results. Someone who wants that behavior can simply do
random.choice(rule34Py.search([tags...])).

So remove the tags functionality from random_post() and make it always
behave like the website's function.

Further, make the rule34Py.random_post_id() method public and correct
its return type to 'int'. There is little reason that it should be
private, when other methods with similar levels of testing and support
are public.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>