pyproject.toml: restrict python version <4.0
The requests-ratelimiter module "requires" python >=3.7,<4.0. Because it
is a dependency of this project, when you try to process the project
dependencies with poetry, you get a satisfaction warning that the python
versions required by this project (>3.9) technically could include
runtimes 4.0+, breaking compat.

It's an odd rqeuirement, that honestly should be fixed in upstream
requests-ratelimiter. Until we can get that upstreamed, just add '<4.0'
to our python requirement to resolve the error.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>