Makefile: clean up rules
The Twine utility is now used to publish the project artifacts, but it
is not included in the pyproject dependencies. Add it to the deps.

Use a small python script to programmatically extract the project name
and version from the pyproject.toml file, since that is the source of
the artifact names.

Generally clean up the Makefile targets.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>