Use poetry as the centralized env and build tool
The project currently uses setuptools, build, and twine to coordinate
project build and publishing steps. And it leaves the business of
managing the python environment as an exercise to the user.

Further, the current setuptools pyproject implementation does not
correctly handle Maintainership contact information, and the python
dependencies are slightly messed up.

We should instead use poetry - which is a modern build tool for python
that most of our dependencies use - to manage the python venv and
coordinate dependencies.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>