Remove module metadata and rule34Py.version()
The modern (post-PEP440) standard for distributing package metadata is
to make use of the package PKG-INFO file and the importlib.metadata
standard module.

Remove the obsolete module-level __author__ and __version__ dunders,
along with the rule34Py.version() method, in favor of the package
metadata.

Add a new guide, explaining to users how to extract the metadata.

Signed-off-by: Riparian Commit <ripariancommit@protonmail.com>